requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
black==23.3.0
boto3==1.4.4
botocore==1.5.95
cattrs==23.1.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
daphne==4.0.0
defusedxml==0.7.1
dj-database-url==2.0.0
Django==4.2.2
django-crispy-forms==2.0
django-schema-graph==3.1.0
django-storages==1.13.2
django-templated-mail==1.1.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
docutils==0.20.1
et-xmlfile==1.1.0
exceptiongroup==1.1.1
fonttools==4.40.0
fpdf==1.7.2
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
jmespath==0.10.0
kiwisolver==1.4.4
matplotlib==3.7.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.25.0
oauthlib==3.2.2
openai==0.27.8
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pathspec==0.11.1
Pillow==10.0.0
platformdirs==3.6.0
psycopg2-binary==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
pyOpenSSL==23.2.0
pyparsing==3.1.0
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.1.13
service-identity==23.1.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
tomli==2.0.1
tqdm==4.65.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
uvicorn==0.17.3
whitenoise==6.5.0
XlsxWriter==3.1.2
yarl==1.9.2
zope.interface==6.0
